checkpoint Issue #43: className based registration only
prior to this commit, the registry was maintained with a projectName
index and className index.

The project name index is almost useless as different names can be used
in different configurations for the same configuration. the only
meaningful name is the className of the configuration.

While adding determineCurrentVersion tests I ran smack dab into an
ambiguity caused by a configuration with two names. I had been
considering the change for awhile now, so with this concrete problem in
hand, it was time to pull the trigger.

Switched registration to use a configurationRegistry and a
baselineRegistry using the className for keys. I split into two
registries to make lookup for baseline or configuration simpler ... the
baseline className key and configuration className key both reference
the same registration object.

This refactoring simplified the registration logic quite a bit and for
now I'm throwing errors under certain suspicious circumstances.

still working through test conversion....