checkpoint Issue #43 ... refine registrationsCompareEqual:
MetacelloDetermineCurrentVersionForLoadScriptTest>>test004new exposed a
hasConflicts: bug that caused me to rethink the definition of
registrationsCompareEqual: and I've decided that className and
versionString and operator are the only things worth conflicting on ...
all other changes repository, pre/post doits, load list, etc. are valid
things to change and should be honored without question. Of course
versionString is the big conflict, but typically that leads to an
onUpgrade: or onDowngrade: clause being triggered.

Also cleaned up some crift that was still lying around