Issue #68: looks like `specListDetect:` is the *correct* method
In debugging the test, I found that I had already created protocol for
"doing the right thing".

If you remember I had mentioned that `packageNamed:` was not the right
message to use? Well, it turns out that `detect:` isn't the right set of
messages either.

If you inspect the packages iv in the versionSpec, you'll see that
you've got a *MetacelloPackagesSpec* and if you poke around in
*MetacelloPackagesSpec* you'll see that it has a `list` iv and a
`memberMap`. This is where things get odd.

The `list` iv **is** the collection that needs to be accessed for doing
editting ... the `memberMap` iv is a condensed version of the specs in
the `list` and are copies, so any changes in them aren't preserved.

When I first started Metacello, it was natural to use `do:` and friends
to access the `memberMap`, because at the time, there was no direct
editting of the `list`...

When I introduced the toolbox, I needed to directly access the specs in
the `list` and I created the `specList*` series of messages ...

Given the fact that the toolbox exposes folks directly to the `list`, I
suppose we need to consider switching to the standard collection
accessing messages... Of course, if I look at things a bit closer, we
may find that users are also directly exposed to the collection api for
the `memberMap` variant....hopefully not:)