Issue #87: `flushForScriptGet` causes GitHub repo to flush the downloadCache
`flushForScriptGet` is sent during `get` command to ensure that a fresh
download is attempted from GitHub before loading up the configuration or
baseline.

MetacelloScriptGitHubDownloadNotification is simply there to give the
test something to look for ...

remove questionable MetacelloMCBaselineOfProjectSpec>>projectPackage was
setting projectVersion from baseline, but I believe that should have
obsoleted a while ago ... the repository description for the baseline is
the only way to define the commitish values.