update docs to reflect convention of using name 'packages'.
We have quite a few different repositories, and a repository inside a
repository etc. What's really in the filetree is packages, so I
personally prefer to have that as the convention.

Note it's just a convention though, the directory can have any name but
it usually helps if people know what to expect.