squeak: Fix cache preferences for BitBucket and GitHub Fetch repos
During the migration from the filetree-only git repos to the more general ones we did not copy the necessary accessors for the settings to be saved and read correctly. Further, we used ClassVars instead of class-side InstVars, which meant that BitBucket and GitHub always shared the same folder. This is now repaired.

I dare claim that a migration script is not required here since setting the preference had been broken anyways, so there's good reason to assume that people left it at the default thus far.