Fixes precedence of Squeak's platform attributes.
In particular, the alpha-fallback got appended only after the attribute for the current system version, which is wrong because a more recent symbolic version could never be matched. Also increase the latest known Squeak 5 to version 5.3 to make that alpha-fallback work properly again in configurations that do not have a squeak6.x configured yet for 6.0alpha, which is the current Trunk.